Add ParseSchema element to Schema enumeration.
This adds the ability to perform arbitrary parsing and/or validation
in the process of evaluation of a schema. Because this introduces a
failure case that must be known at the time of schema construction,
it adds an error type parameter to the schema. In order to preserve
the contract for automatic exemplar generation, we use an uninhabited
type as the value of this parameter when passing schema to the
'generate' function.